// shared/types.ts
export type Face = {
  acceptability: number;
  images: Images;
  landmarks?: Landmarks;
  bounding_box: BoundingBox;
  mask?: number;
  pitch: number;
  quality: number;
  roll: number;
  yaw: number;
  tracking_id?: string;
};

export type Point = { x: number; y: number };
type BoundingBox = { origin: Point };

export type Images = {
  expanded?: string;
  recognition_input_image?: string;
};

export type Landmarks = {
  left_eye: Point;
  right_eye: Point;
  left_mouth: Point;
  right_mouth: Point;
  nose: Point;
};
// ...
export type DetectedResult = {
  faces: Face[];
  metadata: {
    width: number;
    height: number;
    codec: string;
    frame_rate: string;
  };
  timestamp: Date;
};
// ...
export type CameraData = {
  //base cam settings
  id: number | undefined;
  name: string;
  direction: number; //Entrance, Exit, Corridor, Other
  rtsp_url: string; //top level rtsp url
  min_match: number | null; //anything below this confidence is not a match
  user_name: string; //TODO: user and password can be set at camera create but shouldn't be available when getting info.
  password: string; //TOOD: shouldn't use passwords here.
  fr_stream_settings: FRStreamSettings | undefined; //how we build recognition stream
  rtsp_stream_info: RTSPStreamInfo | undefined; //how we build the display
  proxy_url: string | undefined; //the base address of the cam proxy server (could also route through rev. proxy
  feed_position: number | undefined;
  //TODO: think. is this really the best place for this? not part of the camera settings
  //state: CameraState | undefined;

  /*is_updating: boolean | undefined;
  is_streaming: boolean | undefined; //TODO: deprecate
  is_detecting: boolean | undefined; //replaces streaming. more precise. is the fr face detection running on this camera stream?

   */
};


//RTSPVideo Api types
export type RTSPChannel = {
  name: string;
  url: string;
  on_demand: boolean;
  debug: boolean;
  status: number;
};

export type RTSPStreamInfo = {
  id: string | undefined;
  name: string;
  channels: Record<string, RTSPChannel>;
};

export type Payload = Record<string, RTSPStreamInfo>;

export type GetRTSPStreamsResp = {
  status: number;
  payload: Payload;
};

//the status of all running camera streams being analyzed for faces
//the result of a fetch to the paravision streams endpoint
//Config information about how paravision is analyzing a camera stream

export type  FRStreamSettings = {
  detect_frame_rate: number;
  detect_mask: boolean;
  enable_tracking: boolean;
  expanded_image_scale: number;
  face_acceptability_tracking_threshold: number | undefined;
  face_quality_tracking_threshold: number | undefined;
  face_similarity_tracking_threshold: number | undefined;
  inference_worker_id: number | undefined;
  max_frames_per_track: number;
  metadata: Metadata | undefined;
  min_frames_per_track: number;
  name: string;
  output_faces_only: boolean | undefined;
  postprocess_worker_id: number | undefined;
  rotation: number;
  skip_identical_frames: boolean;
  source: string;
  tracking_duration: number;
  tracking_min_face_size: number;
};

//pv specific types
export type Metadata = {
  codec: string;
  frame_rate: string;
  height: number;
  width: number;
};